requirements.txt
asgiref==3.3.1
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.1.3
django-cors-headers==3.8.0
djangorestframework==3.12.4
google-api-core==2.0.0
google-api-python-client==2.17.0
google-auth==2.0.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.5
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
idna==3.2
oauthlib==3.1.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytz==2020.4
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.3.0
